Clean up unused imports in unit tests
