+++ test/units/executor/module_common/conftest.py
@@ -1,0 +1,10 @@
+import pytest
+
+
+@pytest.fixture
+def templar():
+    class FakeTemplar:
+        def template(self, template_string, *args, **kwargs):
+            return template_string
+
+    return FakeTemplar()
