+++ test/units/module_utils/basic/test_sanitize_keys.py
@@ -6,7 +6,6 @@
 from __future__ import (absolute_import, division, print_function)
 __metaclass__ = type
 
-import pytest
 from ansible.module_utils.basic import sanitize_keys
