+++ test/units/module_utils/common/validation/test_check_missing_parameters.py
@@ -9,7 +9,6 @@ __metaclass__ = type
 import pytest
 
 from ansible.module_utils._text import to_native
-from ansible.module_utils.common.validation import check_required_one_of
 from ansible.module_utils.common.validation import check_missing_parameters
