+++ test/units/module_utils/urls/test_fetch_file.py
@@ -10,7 +10,6 @@ import os
 from ansible.module_utils.urls import fetch_file
 
 import pytest
-from units.compat.mock import MagicMock
 
 
 class FakeTemporaryFile:
