+++ test/units/modules/test_async_wrapper.py
@@ -9,12 +9,7 @@ import json
 import shutil
 import tempfile
 
-import pytest
-
-from units.compat.mock import patch, MagicMock
 from ansible.modules import async_wrapper
-
-from pprint import pprint
 
 
 class TestAsyncWrapper:
