+++ test/units/plugins/filter/test_core.py
@@ -3,8 +3,6 @@
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 
 from __future__ import absolute_import, division, print_function
-from jinja2.runtime import Undefined
-from jinja2.exceptions import UndefinedError
 __metaclass__ = type
 
 import pytest
