+++ test/units/utils/collection_loader/fixtures/collections/ansible_collections/testns/testcoll/plugins/action/my_action.py
@@ -1,7 +1,7 @@
 from __future__ import (absolute_import, division, print_function)
 __metaclass__ = type
 
-from ..module_utils.my_util import question
+from ..module_utils.my_util import question  # pylint: disable=unused-import
 
 
 def action_code():
