+++ test/units/utils/collection_loader/fixtures/collections/ansible_collections/testns/testcoll/plugins/module_utils/my_other_util.py
@@ -1,4 +1,4 @@
 from __future__ import (absolute_import, division, print_function)
 __metaclass__ = type
 
-from .my_util import question
+from .my_util import question  # pylint: disable=unused-import
